refactors helpmenu once more.
get rid of the idiotic "number of items" picker in the settings form,
just render inputs for all ten available slots.
rework the output component accordingly.

improves the labelin of corresponding form elements in the settings form
while at it.
