+++ lang/en/theme_ucsf.php
@@ -64,20 +64,18 @@ $string['end_monthlydesc'] = 'Set time when you want your monthly alert to end.'
 $string['end_weeklydesc'] = 'Set time when you want your weekly alert to start/end.';
 $string['enddate'] = 'End date';
 $string['generalsettings'] = 'General settings';
-$string['helpfeedback'] = 'Help/Feedback';
-$string['helpfeedbackenabled'] = 'Enable Help/Feedback menu';
-$string['helpfeedbackenableddesc'] = ' Click to enable Help/Feedback menu';
-$string['helpfeedbacklink'] = 'Help/Feedback link to a static page';
-$string['helpfeedbacklinkdesc'] = 'Help/Feedback link to a static page. Example: http://www.ucsf.edu/static/help.html';
-$string['helpfeedbacklinklabel'] = "Enter label name";
-$string['helpfeedbacklinklabeldesc'] = "Custom label name";
-$string['helpfeedbacklinktarget'] = 'Open Help/Feedback link in new window';
-$string['helpfeedbacklinktargetdesc'] = 'Should Help/Feedback link opens in a new window';
-$string['helpfeedbackno'] = 'Help/Feedback {$a->help}';
-$string['helpfeedbacknumlinks'] = 'Select number of links';
-$string['helpfeedbacksettings'] = 'Help/Feedback settings';
-$string['helpfeedbacksubsectiontitle'] = 'Help and Feedback';
-$string['helpmenu'] = 'Help menu';
+$string['helpfeedback'] = 'Helpmenu';
+$string['helpfeedbackenabled'] = 'Enabled';
+$string['helpfeedbackenableddesc'] = ' Check to enable the display of the helpmenu in the site navigation banner.';
+$string['helpfeedbacklink'] = 'Link';
+$string['helpfeedbacklinkdesc'] = 'The URL that this menu item links to. Leave this field empty to exclude this item from the menu.';
+$string['helpfeedbacklinklabel'] = "Title";
+$string['helpfeedbacklinklabeldesc'] = "The menu item title.";
+$string['helpfeedbacklinktarget'] = 'Open link in new window';
+$string['helpfeedbacklinktargetdesc'] = "Check to have this menu item's link open up in a new window.";
+$string['helpfeedbackno'] = 'Menu item {$a}';
+$string['helpfeedbacksettings'] = 'Helpmenu';
+$string['helpmenu'] = 'Helpmenu';
 $string['invalidenddate'] = 'Invalid end date.';
 $string['invalidstartandenddate'] = 'Invalid start date and end date.';
 $string['invalidstartdate'] = 'Invalid start date.';
